fix float in lr

+++ train/train.py
@@ -150,7 +150,7 @@ if __name__ == "__main__":
     parser.add_argument('--checkpoint', type=str, default='')
     parser.add_argument('--model', type=str, default='t-lite', choices=['t-lite', 'qwen'])
     parser.add_argument('--epochs', type=int, default=10)
-    parser.add_argument('--lr', type=int, default=3e-5)
+    parser.add_argument('--lr', type=float, default=3e-5)
     parser.add_argument('--batch_size', type=int, default=32)
     parser.add_argument('--save_steps', type=int, default=2000)
     parser.add_argument('--log_steps', type=int, default=100)
